+++ src/Loader/ShadersWGSL/mmdOutline.fragment.ts
@@ -10,7 +10,7 @@ const Shader = /* wgsl */`
 uniform color: vec4f;
 
 #ifdef ALPHATEST
-varying vUV: vec2;
+varying vUV: vec2f;
 var diffuseSamplerSampler: sampler;
 var diffuseSampler: texture_2d<f32>;
 #endif